storage-initializer: extend the entrypoint to also decrypt the model
In the Confidential Containers use case, the model gets encrypted
before being uploaded in the remote storage (s3 bucket for example).
The existing entrypoint simply downloads whatever it is present
at the specified path in the data connection, but if the model
file/folder is encrypted, the actual container that then runs
the model will fail, either because it doesn't find it (name doesn't
match), or because it is encrypted.

Extend the initializer-entrypoint by adding a wrapper that simply call the
original initializer-entrypoint and then tries to fetch the key from
a remote attester to decrypt the model.
There is no need to specify or hardcode the remote attester (Trustee) location
since the agents inside the Confidential VM running the container already
know the remote ip, so pinging them via localhost is more than enough.
The whole attestation part is taken care by the Confidential Containers
operator, so it is outside the scope of the storage initializer to understand
the attestation flow.

All this init container has to do is:
1. download the model (initializer-entrypoint original purpose)
2. decrypt the model by fetching a given key from an unit listening at localhost
3. if the model is a TAR, extract it in the same folder where it was downloaded

The new storage initializer supports both encrypted model files
(for example model.onnx -> model.onnx.enc) and encrypted tar folders
(flan-t5-small/ -> flan-t5-small.tar.gz -> flan-t5-small.tar.gz.enc).

The encrypted model will be searched in $2 (destination path of
initializer-entrypoint python script) and $2/1 (since some models
like .onnx use the additional /1 prefix).

There are a couple of env variable to define the wrapper behavior:
MODEL_ENCRYPTED = define if the model is encrypted or not. When false,
the wrapper doesn't do anything
MODEL_DECRYPTION_KEY = path to the key stored in the remote attester
MODEL_NAME = name of the model file/folder
MODEL_FORMAT = extension of the model, without leading dot (onnx, tar.gz, ...).
If the extension is tar.gz, the wrapper script will try to unpack the folder
after it has been decrypted.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>